[Edit] RecordViewModel 리팩토링 #142
- WrittenUser 모델 변경
- Drink 모델 오타 수정
- 기존 RecordViewModel -> ExRecordViewModel 변경
- RecordVieModel 새로 생성 및 코드 리팩토링
- FireStroageService 의 uploadImageToStorage() 수정
