--- JUDA/ViewModel/Posts/RecordViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  RecordService.swift
-//  JUDA
-//
-//  Created by Minjae Kim on 3/8/24.
-//
-
-import Foundation
